extracted UnityMeshData from SceneImporter
